recommendation_engine/requirements.txt
dnspython==2.2.1
joblib==1.1.0
numpy==1.22.4
pandas==1.4.2
pymongo==4.1.1
python-dateutil==2.8.2
pytz==2022.1
scikit-surprise==1.1.1
scipy==1.8.1
six==1.16.0
